+++ lib/Compilation.js
@@ -104,21 +104,6 @@ const { isSourceEqual } = require("./util/source");
 /** @typedef {import("./DependenciesBlock")} DependenciesBlock */
 /** @typedef {import("./Dependency").DependencyLocation} DependencyLocation */
 /** @typedef {import("./Dependency").ReferencedExports} ReferencedExports */
-/** @typedef {import("./LazyBarrel").DependencyGroup} LazyBarrelGroup */
-/**
- * Defines the lazy barrel state type used by this module.
- * @typedef {object} LazyBarrelState
- * @property {Set<string> | true} forwardedIds export names requested so far, true for all
- * @property {LazyBarrel | undefined} lazyBarrelInfo deferred dependencies, undefined until classified
- */
-/**
- * Defines the lazy barrel unlazy item type used by this module.
- * @typedef {object} LazyBarrelItem
- * @property {ModuleFactory} factory factory for the request
- * @property {Dependency[]} dependencies deferred dependencies of the request
- * @property {string | undefined} context request context
- * @property {Module} originModule the lazy barrel module
- */
 /** @typedef {import("./Entrypoint").EntryOptions} EntryOptions */
 /** @typedef {import("./Module").NameForCondition} NameForCondition */
 /** @typedef {import("./Module").BuildInfo} BuildInfo */
@@ -1385,12 +1370,12 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 		/** @type {UnsafeCachePredicate} */
 		this._unsafeCachePredicate =
 			typeof unsafeCache === "function" ? unsafeCache : () => true;
+
 		/**
-		 * Per side-effect-free module lazy barrel state.
 		 * @private
-		 * @type {WeakMap<Module, LazyBarrelState>}
+		 * @type {LazyBarrel}
 		 */
-		this._lazyBarrelModules = new WeakMap();
+		this._lazyBarrel = new LazyBarrel(this);
 	}
 
 	getStats() {
@@ -1801,7 +1786,7 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 		);
 
 		// lazy barrel: defer re-export targets of side-effect-free modules
-		const hasLazyBarrel = this._classifyLazyBarrelDependencies(module);
+		const hasLazyBarrel = this._lazyBarrel.classify(module);
 
 		/** @type {DependenciesBlock} */
 		let currentBlock;
@@ -1884,10 +1869,13 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 			this.moduleGraph.setParents(dep, currentBlock, module, index);
 			if (
 				hasLazyBarrel &&
-				/** @type {{ _lazyMake?: boolean }} */ (dep)._lazyMake === true
+				// `isLazy` may be missing on custom dependency types not extending the base Dependency
+				typeof dep.isLazy === "function" &&
+				dep.isLazy()
 			) {
 				return;
 			}
+
 			if (this._unsafeCache) {
 				try {
 					const unsafeCachedModule = unsafeCacheDependencies.get(dep);
@@ -1901,7 +1889,7 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 								dep,
 								unsafeCachedModule
 							);
-							this._unlazyBarrelForDependency(
+							this._lazyBarrel.unlazyForDependency(
 								unsafeCachedModule,
 								dep,
 								sortedDependencies
@@ -1919,7 +1907,7 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 								dep,
 								cachedModule
 							);
-							this._unlazyBarrelForDependency(
+							this._lazyBarrel.unlazyForDependency(
 								cachedModule,
 								dep,
 								sortedDependencies
@@ -1984,7 +1972,7 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 									dep,
 									cachedModule
 								); // a3
-								this._unlazyBarrelForDependency(
+								this._lazyBarrel.unlazyForDependency(
 									cachedModule,
 									dep,
 									sortedDependencies
@@ -2124,127 +2112,6 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 		}
 
 		if (--inProgressSorting === 0) onDependenciesSorted();
-	}
-
-	/**
-	 * Classifies the re-export dependencies of a side-effect-free module (lazy barrel)
-	 * into deferrable groups and marks the deferred ones.
-	 * @private
-	 * @param {Module} module the module whose dependencies are processed
-	 * @returns {boolean} true, when some dependencies were deferred
-	 */
-	_classifyLazyBarrelDependencies(module) {
-		const lazyBarrelModules = this._lazyBarrelModules;
-		const factoryMeta = module.factoryMeta;
-		if (factoryMeta === undefined || !factoryMeta.sideEffectFree) {
-			return false;
-		}
-		/** @type {LazyBarrel | undefined} */
-		let info;
-		for (const dep of module.dependencies) {
-			const until = dep.getLazyUntil();
-			if (until === null) continue;
-			if (info === undefined) info = new LazyBarrel();
-			if (typeof until === "object" && "local" in until) {
-				info.addTerminal(until.local);
-				continue;
-			}
-			const resourceIdent = dep.getResourceIdentifier();
-			if (resourceIdent === null) continue;
-			const factory = this.dependencyFactories.get(
-				/** @type {DependencyConstructor} */ (dep.constructor)
-			);
-			if (factory === undefined) continue;
-			const category = dep.category;
-			// Match the request grouping key of `processDependencyForResolving`
-			const requestKey =
-				category === esmDependencyCategory
-					? resourceIdent
-					: `${category}${resourceIdent}`;
-			info.addLazy(requestKey, dep, factory, dep.getContext());
-			if (typeof until === "object") {
-				info.addForwardId(until.id, requestKey);
-			} else if (until === Dependency.LAZY_UNTIL_FALLBACK) {
-				info.addFallback(requestKey);
-			}
-		}
-		const state = lazyBarrelModules.get(module);
-		if (info === undefined || info.isEmpty()) {
-			if (state !== undefined) lazyBarrelModules.delete(module);
-			return false;
-		}
-		if (state !== undefined) {
-			// barrel classified after its targets were already requested: replay the
-			// recorded ids so processDependency un-lazies and builds those targets now
-			state.lazyBarrelInfo = info;
-			info.request(state.forwardedIds);
-			// stay lazy only while some targets remain deferred
-			return !info.isEmpty();
-		}
-
-		lazyBarrelModules.set(module, {
-			forwardedIds: new Set(),
-			lazyBarrelInfo: info
-		});
-
-		return true;
-	}
-
-	/**
-	 * Requests the export names that `dependencies` need from a lazy barrel and
-	 * returns the deferred re-export targets that must be built now. Records the
-	 * request when the barrel's dependencies were not classified yet.
-	 * @private
-	 * @param {Module} module the resolved module
-	 * @param {Dependency[]} dependencies the dependencies that resolved to the module
-	 * @returns {LazyBarrelItem[] | undefined} items to process, if any
-	 */
-	_requestLazyBarrelItems(module, dependencies) {
-		const lazyBarrelModules = this._lazyBarrelModules;
-		const state = lazyBarrelModules.get(module);
-
-		if (state === undefined) {
-			const factoryMeta = module.factoryMeta;
-			if (factoryMeta === undefined || !factoryMeta.sideEffectFree) return;
-			const forwardedIds = LazyBarrel.getForwardedIds(dependencies);
-			if (forwardedIds !== true && forwardedIds.size === 0) return;
-			lazyBarrelModules.set(module, {
-				forwardedIds,
-				lazyBarrelInfo: undefined
-			});
-			return;
-		}
-
-		const forwardedIds = LazyBarrel.getForwardedIds(dependencies);
-		if (forwardedIds !== true && forwardedIds.size === 0) return;
-		if (state.forwardedIds !== true) {
-			if (forwardedIds === true) state.forwardedIds = true;
-			else for (const id of forwardedIds) state.forwardedIds.add(id);
-		}
-		const info = state.lazyBarrelInfo;
-		// Pending requests will be replayed during `_classifyLazyBarrelDependencies` later
-		if (info === undefined) return;
-		const groups = info.request(forwardedIds);
-		if (groups.length === 0) return;
-		return groups.map((group) => ({
-			factory: group.factory,
-			dependencies: group.dependencies,
-			context: group.context,
-			originModule: module
-		}));
-	}
-
-	/**
-	 * Queues the deferred dependency groups of a lazy barrel requested by a single dependency.
-	 * @private
-	 * @param {Module} module the resolved module of the dependency
-	 * @param {Dependency} dependency the requesting dependency
-	 * @param {{ factory: ModuleFactory, dependencies: Dependency[], context: string | undefined, originModule: Module | null }[]} sortedDependencies item list to append to
-	 */
-	_unlazyBarrelForDependency(module, dependency, sortedDependencies) {
-		const items = this._requestLazyBarrelItems(module, [dependency]);
-		if (items === undefined) return;
-		for (const item of items) sortedDependencies.push(item);
 	}
 
 	/**
@@ -2651,13 +2518,13 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 			// lazy barrel:
 			// 1. first call: barrel module is not classified yet, so this records the requested
 			//    ids and returns undefined; the targets get built below when
-			//    `processModuleDependencies` runs `_classifyLazyBarrelDependencies` and replays them.
+			//    `processModuleDependencies` runs `_lazyBarrel.classify` and replays them.
 			// 2. later call: barrel module is already built, so `processModuleDependencies` won't
 			//    re-walk it; the newly requested targets must be built here.
-			const lazyItems = this._requestLazyBarrelItems(module, dependencies);
+			const unlazyItems = this._lazyBarrel.request(module, dependencies);
 
-			if (lazyItems !== undefined) {
-				let inProgress = lazyItems.length + 1; // +1 = the module's own dep processing
+			if (unlazyItems !== undefined) {
+				let inProgress = unlazyItems.length + 1; // +1 = the module's own dep processing
 				let errored = false;
 				/**
 				 * @param {(WebpackError | null)=} err error
@@ -2671,7 +2538,7 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 					}
 					if (--inProgress === 0) callback(null, module);
 				};
-				for (const item of lazyItems) {
+				for (const item of unlazyItems) {
 					this.handleModuleCreation(item, (err) => {
 						onJobDone(err && this.bail ? err : null);
 					});
