+++ lib/Dependency.js
@@ -268,6 +268,20 @@ class Dependency {
 	}
 
 	/**
+	 * Whether the lazy barrel currently defers creating this dependency's target module (lazy barrel optimization).
+	 * @returns {boolean} true while deferred, so it must not be processed or rendered
+	 */
+	isLazy() {
+		return false;
+	}
+
+	/**
+	 * Sets whether the lazy barrel defers creating this dependency's target module (lazy barrel optimization).
+	 * @param {boolean} value true to defer, false to create it now
+	 */
+	setLazy(value) {}
+
+	/**
 	 * Returns the referenced module and export
 	 * @deprecated
 	 * @param {ModuleGraph} moduleGraph module graph
