+++ lib/dependencies/HarmonyImportDependency.js
@@ -114,6 +114,22 @@ class HarmonyImportDependency extends ModuleDependency {
 	}
 
 	/**
+	 * Whether the lazy barrel currently defers creating this dependency's target module (lazy barrel optimization).
+	 * @returns {boolean} true while deferred, so it must not be processed or rendered
+	 */
+	isLazy() {
+		return this._lazyMake;
+	}
+
+	/**
+	 * Sets whether the lazy barrel defers creating this dependency's target module (lazy barrel optimization).
+	 * @param {boolean} value true to defer, false to create it now
+	 */
+	setLazy(value) {
+		this._lazyMake = value;
+	}
+
+	/**
 	 * Returns true if this dependency can be concatenated
 	 * @returns {boolean} true if this dependency can be concatenated
 	 */
@@ -383,7 +399,7 @@ HarmonyImportDependency.Template = class HarmonyImportDependencyTemplate extends
 
 		const connection = moduleGraph.getConnection(dep);
 		// deferred by lazy barrel: never resolved, must not render a missing module
-		if (connection === undefined && dep._lazyMake) {
+		if (connection === undefined && dep.isLazy()) {
 			return;
 		}
 		if (connection && !connection.isTargetActive(runtime)) return;
